Inlining and cloning stuff
Many issues were raising at compilation mode because uninitializedBody of methods were being stored without mateification. Then, when graal clones, mateification were being lost. I added a hack to the deepcopy but that was bringing to many issues too because there are times were we do not want the mateification. For instance, for the methods that are duplicated without mate nodes for their execution at the meta level. The solution was too just create the uninitialized body mateified, but add the metalevel methods before that mateification.